requirements.txt
DracoPy
gltflib
matplotlib
numpy
Pillow
pyproj
scipy
torch
tqdm
transformers
trimesh